arduino-fsr/pygui/requirements.txt
dearpygui
pyserial
